[EI-128] [web] rename: Rename dialog menu key to dialog key
