+++ web/app/ui/components/numerical-guidance/metadata-list-item.tsx
@@ -4,7 +4,7 @@ import { DotsHorizontalIcon } from '@heroicons/react/solid';
 import { IndicatorBoardMetadataResponse } from '@/app/store/querys/numerical-guidance/indicator-board-metadata.query';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/use-selected-indicator-board-metadata.hook';
 import { useDialogMenu } from '../view/molocule/dialog-menu';
-import { DIALOG_KEY } from '@/app/utils/keys/dialog-menu-key';
+import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 
 type MetadataListItemProps = {
   item: IndicatorBoardMetadataResponse;
