+++ web/app/ui/components/view/molocule/alert-dialog/use-alert-dialog.hook.ts
@@ -1,7 +1,7 @@
 import { useDialogStore } from '@/app/store/stores/dialog.store';
-import { DialogMenuKey } from '@/app/utils/keys/dialog-menu-key';
+import { DialogKey } from '@/app/utils/keys/dialog-key';
 
-export function useAlertDialog(key: DialogMenuKey) {
+export function useAlertDialog(key: DialogKey) {
   const action = useDialogStore((state) => state.action);
   const isOpen = useDialogStore((state) => state.isOpen[key]);
