+++ web/app/ui/components/view/molocule/dialog-menu/dialog-menu-root.tsx
@@ -4,11 +4,11 @@ import { Transition } from '@headlessui/react';
 import { DialogMenuItem } from './dialog-menu-item';
 import { DialogMenuContext } from './dialog-menu.context';
 import { useDialogMenu } from './use-dialog-menu.hook';
-import { DialogMenuKey } from '@/app/utils/keys/dialog-menu-key';
+import { DialogKey } from '@/app/utils/keys/dialog-key';
 import { filterChildrenByType } from '@/app/utils/helper';
 
 type DialogMenuProps = {
-  dialogKey: DialogMenuKey;
+  dialogKey: DialogKey;
 };
 
 const getDialogMenuItems = (children: React.ReactNode) => {
