+++ web/app/ui/components/view/molocule/dialog-menu/use-dialog-menu.hook.ts
@@ -1,8 +1,8 @@
 import { useRef } from 'react';
 import { useDialogStore } from '@/app/store/stores/dialog.store';
-import { DialogMenuKey } from '@/app/utils/keys/dialog-menu-key';
+import { DialogKey } from '@/app/utils/keys/dialog-key';
 
-export function useDialogMenu(key: DialogMenuKey) {
+export function useDialogMenu(key: DialogKey) {
   const action = useDialogStore((state) => state.action);
   const isOpen = useDialogStore((state) => state.isOpen[key]);
   const position = useDialogStore((state) => state.position);
